Test refactoring vol 5 (#402)
* fix exclusions

* fix minor bugs and add tests

* Add update Order Tests testing different corner cases

* refactor order update scenarios

* fix npe condition

+++ goods-partner-backend/src/main/java/com/goodspartner/exception/OrderNotFoundException.java
@@ -4,7 +4,7 @@ public class OrderNotFoundException extends RuntimeException {
 
     private final static String ORDER_NOT_FOUND_MESSAGE = "Order with id: %s not found";
 
-    public OrderNotFoundException(int id) {
+    public OrderNotFoundException(long id) {
         super(String.format(ORDER_NOT_FOUND_MESSAGE, id));
     }
 }
